requirements.txt
pandas
requests
os
fetch_ucrepo
matplotlib.pyplot
hashlib
zipfile
ProfileReport